adjust condition

+++ install.sh
@@ -49,7 +49,9 @@ ${sudoCmd} ${systemPackage} update
 ${sudoCmd} ${systemPackage} install curl wget git jq -y
 
 mkdir -p /usr/local/etc/v2script
-wget -q https://raw.githubusercontent.com/phlinhng/v2ray-tcp-tls-web/beta/config/v2scirpt.json -O /usr/local/etc/v2script/config.json
+if [ ! -f "/usr/local/etc/v2script/config.json" ]; then
+  wget -q https://raw.githubusercontent.com/phlinhng/v2ray-tcp-tls-web/beta/config/v2scirpt.json -O /usr/local/etc/v2script/config.json
+fi
 wget -q https://raw.githubusercontent.com/phlinhng/v2ray-tcp-tls-web/beta/v2script.sh -O /usr/local/bin/v2script
 chmod +x /usr/local/bin/v2script && /usr/local/bin/v2script
